% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinking.R
\docType{package}
\name{drinking}
\alias{drinking}
\alias{drinking-package}
\title{drinking: Get drinking events from movement data}
\description{
drinking provides usefull tools to find drinking events from movement data. 
The main function assumes that when animals cross a certain threshold of
distance to water they go drinking. It identifies unique drinking event given
a certain return time above which it is considered as a new visit.
}

