% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance2water.R
\name{get_distance2water}
\alias{get_distance2water}
\title{Get distance to water}
\usage{
get_distance2water(df, water_layer, projstring)
}
\arguments{
\item{df}{a dataframe with at least coordinates  \code{df$x}, \code{df$y} and
date and time in \code{df$dateTime} in POSIXct class.}

\item{water_layer}{a SpatialPointsDataFrame of waterhole with column 
\code{$name} in data slot.}
}
\description{
\code{get_distance2water} gets distance to water from SpatialPointsDataFrame
water_layer
}
\examples{
get_distance2water(df,water_layer,projstring)
}

