% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drinking_event_function.R
\name{get_drinking_event}
\alias{get_drinking_event}
\title{Get drinking events from movement data}
\usage{
get_drinking_event(df, water_layer = NULL, threshold = 500,
  returnTime = 6, projstring, id = F)
}
\arguments{
\item{df}{a dataframe with at least coordinates  \code{df$x}, \code{df$y} and
date and time in \code{df$dateTime} in POSIXct class.}

\item{water_layer}{a SpatialPointsDataFrame of waterhole with column 
\code{$name} in data slot. If NULL, function assumes that column
\code{df$dist2water} gives distance to water}

\item{threshold}{distance below which animals are considering going to drink}

\item{returnTime}{return time (in hours) above which an animals can be considering
returning drinking to the former waterhole}

\item{projstring}{CRS projection of data}

\item{id}{if TRUE calculate distance for individuals identified by column \code{df$id}}
}
\description{
\code{get_drinking_event} assumes that when animals cross a certain
\code{threshold} of distance to water they go drinking. It identifies unique
drinking events given a certain \code{return} time above which it is
considered as a new visit. Distance to water might be provided as a
SpatialPointsDataFrame or as two columns (named \code{dist2water} & \code{waterholes}) in the
dataframe \code{df}.
}
\examples{
get_drinking_event(df,water_layer=NULL,threshold=500,return=6,origin_julday,projstring)
}

