% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_voronoi.R
\name{voronoipolygons}
\alias{voronoipolygons}
\title{Create Voronoi Polygons from a SpatialPoints}
\usage{
voronoipolygons(layer, extent)
}
\arguments{
\item{layer}{is a SpatialPoints object}

\item{extent}{is an extent object from raster packages to set the extent of calculation of polygons}
}
\value{
a SpatialPolygonsDataFrame with the Voronoi polygons
}
\description{
\code{voronoi} creates voronoi polygons. The functions use the R package
deldir and is modified from Carson Farmer :
http://carsonfarmer.com/2009/09/voronoi-polygons-with-r/
}
\examples{
voronoipolygons(layer,extent)
}

